package classes into different packages A-packages
