+++ src/main/java/duke/command/ByeCommand.java
@@ -1,3 +1,5 @@
+package duke.command;
+
 public class ByeCommand extends Command<String>{
 
     public ByeCommand() {
