+++ src/main/java/duke/command/MarkCommand.java
@@ -1,10 +1,16 @@
-public class UnmarkCommand extends Command<String>{
+package duke.command;
+
+import duke.storage.Storage;
+import duke.task.Task;
+import duke.tasklist.TaskList;
+
+public class MarkCommand extends Command<String>{
 
     private String text;
     private TaskList list;
     private Storage storage;
 
-    public UnmarkCommand(String text, TaskList list, Storage storage) {
+    public MarkCommand(String text, TaskList list, Storage storage) {
         this.text = text;
         this.list = list;
         this.storage = storage;
@@ -16,9 +22,9 @@ public class UnmarkCommand extends Command<String>{
         int taskNumber = intSearch(text) - 1;
         if (taskNumber < list.getSize() && taskNumber >= 0) {
             Task intendedTask = list.getTask(taskNumber);
-            intendedTask.setDone(false);
+            intendedTask.setDone(true);
             storage.writeToFile(list);
-            System.out.println("  " + "AIYO! I've marked this task as not done yet:\n"
+            System.out.println("  " + "SOLID! I've marked this task as done:\n"
                     + "    " + intendedTask);
         } else {
             System.out.println("EH HULLO!! Task does not exist! Check again hehe");
