+++ src/main/java/duke/exception/DukeManager.java
@@ -1,3 +1,7 @@
+package duke.exception;
+
+import duke.exception.DukeException;
+
 public class DukeManager {
 
     public DukeManager() {
