+++ src/main/java/duke/parser/Parser.java
@@ -1,3 +1,11 @@
+package duke.parser;
+
+import duke.exception.DukeException;
+import duke.storage.Storage;
+import duke.tasklist.TaskList;
+import duke.ui.Ui;
+import duke.command.*;
+
 public class Parser {
 
     private String command;
