+++ src/main/java/duke/storage/Storage.java
@@ -1,3 +1,12 @@
+package duke.storage;
+
+import duke.tasklist.TaskList;
+import duke.exception.DukeException;
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.ToDo;
+
 import java.io.*;
 import java.util.ArrayList;
 import java.util.Scanner;
