+++ src/main/java/duke/tasklist/TaskList.java
@@ -1,3 +1,7 @@
+package duke.tasklist;
+
+import duke.task.Task;
+
 import java.util.ArrayList;
 
 public class TaskList {
