+++ src/main/java/duke/ui/Ui.java
@@ -1,3 +1,10 @@
+package duke.ui;
+
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.ToDo;
+import duke.tasklist.TaskList;
+
 import java.util.Scanner;
 
 public class Ui {
